Parse property notes, and consolidate activityTypeSelectors and propertySelectors into just selectors, since they both work\!
