+++ src/types/index.d.ts
@@ -13,7 +13,7 @@ export interface Property {
   name: string;
   id: string;
   url: string;
-  // notes: string;
+  notes: string;
   // domain: ASType;
   // range: ASType;
 }
